api/requirements.txt
flask==2.0.1
kafka-python==2.0.2
werkzeug==2.0.3
